requirements.txt
numpy>=1.26,<3
PyYAML>=6,<7
Pillow>=10,<12
tqdm>=4.66,<5
